column exclusion - properly handle schema changes
